backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
awscli==1.37.25
boto3==1.36.25
botocore==1.36.25
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
docutils==0.16
fastapi==0.115.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jmespath==1.0.1
packaging==24.2
pluggy==1.5.0
pyasn1==0.6.1
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
rsa==4.7.2
s3transfer==0.11.2
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
watchtower==3.3.1
